+++ @types/register/index.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@flemist/test-utils/dist/lib/register/register" />