+++ src/binding/import.cjs
@@ -1,0 +1,4 @@
+'use strict'
+
+// eslint-disable-next-line node/no-missing-require
+module.exports = require('../../build/Release/binding.node')