+++ src/index.ts
@@ -1,0 +1,22 @@
+export {
+  rdtsc,
+  ThreadPriority,
+  ProcessPriority,
+  getThreadPriority,
+  setThreadPriority,
+  getProcessPriority,
+  setProcessPriority,
+  isWin,
+} from './binding'
+
+export {
+  calcPerformance,
+} from './calcPerformance'
+
+export {
+  calcPerformanceAsync,
+} from './calcPerformanceAsync'
+
+export {
+  runInRealtimePriority,
+} from './runInRealtimePriority'